% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk.R
\docType{methods}
\name{API}
\alias{API}
\alias{assay,GenoSet,ANY-method}
\alias{assay<-,GenoSet,ANY,ANY-method}
\alias{assayNames,GenoSet-method}
\alias{assays,GenoSet-method}
\alias{colData,GenoSet-method}
\alias{pData,GenoSet-method}
\title{Converting eSet API to RangedSummarizedExperiment}
\usage{
\S4method{pData}{GenoSet}(object)

\S4method{colData}{GenoSet}(x)

\S4method{assays}{GenoSet}(x, ..., withDimnames = TRUE)

\S4method{assay}{GenoSet,ANY}(x, i)

\S4method{assay}{GenoSet,ANY,ANY}(x, i) <- value

\S4method{assayNames}{GenoSet}(x)
}
\arguments{
\item{object}{GenoSet}

\item{x}{GenoSet}

\item{...}{extra arguments}

\item{withDimnames}{scalar logical, should assays be returned with dimnames? Currently always TRUE for GenoSet.}

\item{i}{name or index of assay to select}

\item{value}{new value to insert in assays, a matrix-like object}
}

