\name{assayDataElement<-}
\alias{assayDataElement<-}
\title{Set assayDataElement, attaching on-disk resource if necessary}
\usage{
  assayDataElement(object, elt, value) <- value
}
\arguments{
  \item{object}{eSet}

  \item{elt}{character, assayDataElement name}

  \item{value}{input data to assayDataElement}
}
\value{
  eSet
}
\description{
  Set assayDataElement, attaching on-disk resource if
  necessary
}
\author{
  Peter M. Haverty \email{phaverty@gene.com}
}

