\name{assayDataElement}
\alias{assayDataElement}
\title{Get assayDataElement, attaching on-disk resource if necessary...}
\usage{assayDataElement(object, elt)}
\description{Get assayDataElement, attaching on-disk resource if necessary}
\details{Get assayDataElement, attaching on-disk resource if necessary}
\value{assayDataElement, matrix, DataFrame, or the like}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{object}{eSet}
\item{elt}{character}}
