% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds2Rle}
\alias{bounds2Rle}
\title{Convert bounding indices into a Rle}
\usage{
bounds2Rle(bounds, values, n)
}
\arguments{
\item{bounds}{matrix, two columns, with first and last index, like from boundingIndicesByChr}

\item{values}{ANY, some value to be associated with each range, like segmented copy number.}

\item{n}{integer, the expected length of the Rle, i.e. the number of features in the
genome/target ranges processed by boundingIndicesByChr.}
}
\value{
Rle
}
\description{
Given a matrix of first/last indices, like from boundingIndicesByChr, and values for
each range, convert to a Rle.  This function takes the expected length of the Rle, n,
so that any portion of the full length not covered by a first/last range will be a
run with the value NA.  This is typical in the case where data is segmented with CBS
and some of the data to be segmented is NA.
}
\seealso{
Other "segmented data": 
\code{\link{rangeSegMeanLength}()},
\code{\link{runCBS}()},
\code{\link{segPairTable}()},
\code{\link{segTable}()},
\code{\link{segs2Granges}()},
\code{\link{segs2RleDataFrame}()},
\code{\link{segs2Rle}()}
}
\concept{"segmented data"}
