% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcGC}
\alias{calcGC}
\title{Calculate GC Percentage in windows}
\usage{
calcGC(object, bsgenome, expand = 1e+06)
}
\arguments{
  \item{object}{GenomicRanges or GenoSet}

  \item{bsgenome}{BSgenome, like Hsapiens from
  BSgenome.Hsapiens.UCSC.hg19}

  \item{expand}{scalar integer, amount to expand each range
  before calculating gc}
}
\value{
numeric vector, fraction of nucleotides that are G or C in
expanded ranges of \code{object}
}
\description{
Local GC content can be used to remove GC artifacts from
copynumber data (see Diskin et al, Nucleic Acids Research,
2008, PMID: 18784189). This function will calculate GC
content fraction in expanded windows around a set of ranges
following example in
http://www.bioconductor.org/help/course-materials/2012/useR2012/Bioconductor-tutorial.pdf.
Currently all ranges are tabulated, later I may do
letterFrequencyInSlidingWindow for big windows and then
match to the nearest.
}
\examples{
\dontrun{ data(genoset) }
\dontrun{ library(BSgenome.Hsapiens.UCSC.hg19) }
\dontrun{ gc = calcGC(genoset.ds, Hsapiens) }
}

