% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{calcGC2}
\alias{calcGC2}
\title{Calculate GC Percentage in sliding window}
\usage{
calcGC2(dna)
}
\arguments{
  \item{dna}{BSgenome or DNAStringSet} \item{window}{scalar
  integer, calculate GC content in a sliding (by one base)
  window of this size.}
}
\value{
SimpleRleList, integer 1e4 * GC fraction, chromosomes 1:22,
X and Y
}
\description{
Local GC content can be used to remove GC artifacts from
copynumber data (see Diskin et al, Nucleic Acids Research,
2008, PMID: 18784189). This function will calculate GC
content fraction in expanded windows around a set of ranges
following example in http://www.bioconductor.org/help/
course-materials/2012/useR2012/Bioconductor-tutorial.pdf.
Values are as.integer( 1e4 * fraction ) for space reasons.
}
\examples{
\dontrun{ library(BSgenome.Hsapiens.UCSC.hg19) }
\dontrun{ gc = calcGC2(Hsapiens) }
}

