% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{chrPartitioning}
\alias{chrPartitioning}
\title{Partitioning by Chromosome}
\usage{
chrPartitioning(object)
}
\arguments{
\item{object}{GenoSet or GenomicRanges}
}
\value{
PartitioningByEnd
}
\description{
Get indices of first and last element in each chromosome.
}
