% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{cn2lr}
\alias{cn2lr}
\alias{cn2lr,DataFrame-method}
\alias{cn2lr,matrix-method}
\alias{cn2lr,numeric-method}
\title{Take vector or matrix of copynumber values, convert to log2ratios}
\usage{
cn2lr(x, ploidy)

\S4method{cn2lr}{numeric}(x, ploidy)

\S4method{cn2lr}{matrix}(x, ploidy)

\S4method{cn2lr}{DataFrame}(x, ploidy)
}
\arguments{
  \item{x}{numeric vector or matrix, or DataFrame with
  numeric-like columns (Rle typicaly). Assumed to be in
  copynumber units.} \item{ploidy}{numeric, of length ncol(x).
  Ploidy of each sample.}
}
\value{
data of same type as "x" transformed into log2ratio units
}
\description{
Utility function for converting copynumber units (2 is
normal) to log2ratio units (two is normal). If ploidy is
provided lr is log2(cn/ploidy), otherwise log2(cn/2).
}
\seealso{
lr2cn
}

