\name{cn2lr}
\alias{cn2lr}
\title{Take vector or matrix of copynumber values, convert to log2ratios
Utility function for converting copynumber units (2 is normal) to log2ratio units (two is normal)}
\usage{
  cn2lr(x)
}
\arguments{
  \item{x}{numeric data in copynumber units}
}
\value{
  data of same type as "x" transformed into log2ratio units
}
\description{
  Take vector or matrix of copynumber values, convert to
  log2ratios Utility function for converting copynumber
  units (2 is normal) to log2ratio units (two is normal)
}
\author{
  Peter M. Haverty \email{phaverty@gene.com}
}
\seealso{
  lr2cn
}

