% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{dim,GenoSet-method}
\alias{dim,GenoSet-method}
\title{Dimensions}
\usage{
\S4method{dim}{GenoSet}(x)
}
\arguments{
  \item{x}{GenoSet}
}
\value{
integer
}
\description{
Dimensions
}

