% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeSummaries.R
\name{.do_rledf_range_summary}
\alias{.do_rledf_range_summary}
\title{Calculate min/max/sum/mean/whichmin/whichmax over ranges on each column of an RleDataFrame.}
\usage{
.do_rledf_range_summary(
  x,
  bounds,
  na.rm = FALSE,
  simplify = TRUE,
  RLEFUN,
  FUN.TYPE = c("numeric", "double", "integer", "logical")
)
}
\arguments{
\item{x}{RleDataFrame}

\item{bounds}{IRanges or matrix, views on every Rle. If \code{bounds} is a matrix, the first
two columns are used as as the starts and stops. Names for the ranges are taken from rownames of the matrix. Such a matrix can be
constructed with \code{boundingIndicesByChr}.}

\item{na.rm}{scalar logical, ignore NAs in calculations?}

\item{simplify}{scalar logical, simplify result? For a single view, a vector, otherwise a matrix with one row per view.}

\item{RLEFUN}{function, internal rle view summary function like .rle_range_means}

\item{FUN.TYPE}{scalar character, the storage mode for the returned vector or matrix (when simplify==TRUE).}
}
\value{
With \code{simplify == TRUE}, a vector for single view or a matrix
otherwise. When \code{simplify == FALSE}, a list of vectors length ncol(x) where each element is of length \code{nrows(bounds)}.
}
\description{
Loop over the Rle objects in an RleDataFrame, calculate the appropriate statistic for each view. If simplify == FALSE,
this function returns a vector for each Rle. If simplify == TRUE, it returns a vector for the case of a single range, otherwise,
a matrix. Rownames for the matrix are taken from the names of the argument \code{bounds}.
}
\seealso{
RleDataFrame boundingIndicesByChr

Other views: 
\code{\link{.do_rledf_views}()}
}
\concept{views}
\keyword{internal}
