\docType{methods}
\name{elementLengths,GenoSet-method}
\alias{elementLengths,GenoSet-method}
\title{ElementLengths for chromosome}
\usage{
\S4method{elementLengths}{GenoSet}(x)
}
\arguments{
  \item{x}{GenoSet}
}
\value{
character
}
\description{
Get elementLengths from locData slot
}
\details{
Get elementLengths from locData slot
}

