% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{elementNROWS,GenoSet-method}
\alias{elementNROWS,GenoSet-method}
\alias{elementNROWS,GenomicRanges-method}
\title{Get elementNROWS from locData slot}
\usage{
\S4method{elementNROWS}{GenoSet}(x)

\S4method{elementNROWS}{GenomicRanges}(x)
}
\arguments{
\item{x}{GenoSet}
}
\value{
character
}
\description{
Get elementNROWS from locData slot
}
