% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fixSegNAs}
\alias{fixSegNAs}
\title{Fix NA runs in a Rle}
\usage{
fixSegNAs(x, max.na.run = 3)
}
\arguments{
  \item{x}{Rle to be fixed}

  \item{max.na.run}{integer, longest run of NAs that will
  be fixed}
}
\value{
Rle
}
\description{
Fix NA runs in a Rle when the adjacent runs have equal
values
}

