% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{genome,GenoSet-method}
\alias{genome,GenoSet-method}
\alias{genome<-,GenoSet-method}
\title{Genome version}
\usage{
\S4method{genome}{GenoSet}(x)

\S4method{genome}{GenoSet}(x) <- value
}
\arguments{
  \item{x}{GenoSet}

  \item{value}{ Typically a \link{Seqinfo} object for the
  \code{seqinfo} setter.

  Either a named or unnamed character vector for the
  \code{seqlevels} setter.

  A vector containing the sequence information to store for
  the other setters.  }
}
\value{
character, e.g. hg19
}
\description{
The genome positions of the features in locData. The UCSC
notation (e.g. hg18, hg19, etc.) should be used.
}
\examples{
data(genoset)
  genome(genoset.ds)
  genome(genoset.ds) = "hg19"
}

