% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{genomeAxis}
\alias{genomeAxis}
\title{Label axis with base pair units}
\usage{
genomeAxis(locs = NULL, side = 1, log = FALSE, do.other.side = TRUE)
}
\arguments{
  \item{locs}{GenomicRanges to be used to draw chromosome
  boundaries, if necessary.  Usually locData slot from a
  GenoSet.}

  \item{side}{integer side of plot to put axis}

  \item{log}{logical Is axis logged?}

  \item{do.other.side}{logical, label non-genome side with
  data values at tick marks?}
}
\value{
nothing
}
\description{
Label an axis with base positions
}
\details{
Label a plot with Mb, kb, bp as appropriate, using tick
locations from axTicks
}
\examples{
data(genoset)
  genoPlot(genoPos(genoset.ds), genoset.ds[,1, "baf"])
  genomeAxis( locs=locData(genoset.ds) )  # Add chromosome names and boundaries to a plot assuming genome along x-axis
  genomeAxis( locs=locData(genoset.ds), do.other.side=FALSE ) # As above, but do not label y-axis with data values at tickmarks
  genomeAxis()           # Add nucleotide position in sensible units assuming genome along x-axis
}
\seealso{
Other "genome plots": \code{\link{genoPlot}},
\code{\link{genoPlot,GenoSetOrGenomicRanges,ANY-method}},
\code{\link{genoPlot,numeric,Rle-method}},
\code{\link{genoPlot,numeric,numeric-method}}
}

