\name{genoset}
\docType{data}
\alias{genoset}
\alias{fake.lrr}
\alias{fake.baf}
\alias{fake.pData}
\alias{genoset.ds}
\alias{baf.ds}
\alias{cn.ds}
\alias{locData.rd}
\title{Example GenoSet, BAFSet, and CNSet objects and the data to create them.}
\description{Fake LRR, BAF, pData and location data were generated and saved as
  fake.lrr, fake.baf, fake.pData and locData.rd. These were used to
  construct the objects genoset.ds, baf.ds, and cn.ds}
\usage{genoset}
\format{A vector containing 141 observations.}
\source{Fake data generated using rnorm and the like.}
\keyword{datasets}
