% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[,GenoSet,ANY-method}
\alias{[,GenoSet,ANY-method}
\alias{[,GenoSet,GenomicRanges-method}
\alias{[,GenoSet,character-method}
\alias{[<-,GenoSet,ANY,ANY,ANY-method}
\title{Subset a GenoSet}
\usage{
\S4method{[}{GenoSet,ANY}(x, i, j, k, ..., drop = FALSE)

\S4method{[}{GenoSet,character}(x, i, j, ..., drop = FALSE)

\S4method{[}{GenoSet,GenomicRanges}(x, i, j, ..., drop = FALSE)

\S4method{[}{GenoSet,ANY,ANY,ANY}(x, i, j, k) <- value
}
\arguments{
  \item{x}{GenoSet}

  \item{i}{character, GRanges, logical, integer}

  \item{j}{character, GRanges, logical, integer}

  \item{k}{character or integer}

  \item{drop}{logical drop levels of space factor?}

  \item{...}{additional subsetting args}

  \item{value}{typically an array-like \R object of a
  similar class as \code{x}.}
}
\description{
Subset a GenoSet
}
\examples{
data(genoset)
  genoset.ds[1:5,2:3]  # first five probes and samples 2 and 3
  genoset.ds[ , "K"]  # Sample called K
  gr = GRanges(ranges=IRanges(start=seq(from=15e6,by=1e6,length=7),width=1,names=letters[8:14]),seqnames=rep("chr17",7))
  genoset.ds[ gr, "K" ]  # sample K and probes overlapping those in rd, which overlap specifed ranges on chr17
}

