\name{lr2cn}
\alias{lr2cn}
\title{Take vector or matrix of log2 ratios, convert to copynumber
Utility function for converting log2ratio units (zero is normal) to copynumber units (two is normal)}
\usage{
lr2cn(x)
}
\arguments{
  \item{x}{numeric data in log2ratio values}
}
\value{
data of same type as "x" transformed into copynumber units
}
\description{
Take vector or matrix of log2 ratios, convert to copynumber
Utility function for converting log2ratio units (zero is
normal) to copynumber units (two is normal)
}
\author{
Peter M. Haverty \email{phaverty@gene.com}
}
\seealso{
cn2lr
}

