% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{names,GenoSet-method}
\alias{names,GenoSet-method}
\title{Get data matrix names}
\usage{
\S4method{names}{GenoSet}(x)
}
\arguments{
  \item{x}{GenoSet}
}
\value{
character
}
\description{
Get names of data matrices. For the time being, this is
\code{assayDataElementNames}. This function used to do
\code{chrNames}.
}

