% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rangeSummaries.R
\name{numCallable}
\alias{numCallable}
\title{Count Rle positions >= min}
\usage{
numCallable(rle, bounds, min)
}
\arguments{
  \item{rle}{integer Rle, no NAs} \item{bounds}{IRanges or
  matrix, positions in Rle to consider. If \code{bounds} is
  a matrix, the first two columns are used as start and end.}
  \item{min}{scalar integer, count Rle positions >= this
  value.}
}
\value{
integer vector of length nrow(bounds)
}
\description{
For Rle coverage vector, count number of positions where
value >= min, think callable bases.
}

