% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{pos,GenoSetOrGenomicRanges-method}
\alias{pos,GenoSetOrGenomicRanges-method}
\title{Chromosome position of features}
\usage{
\S4method{pos}{GenoSetOrGenomicRanges}(x)
}
\arguments{
\item{x}{GRanges GenoSet}
}
\value{
numeric vector of feature positions within a chromosome
}
\description{
Get chromosome position of features/ranges. Defined as floor of mean of start and end.
}
\examples{
  data(genoset,package="genoset")
  pos(genoset.ds)  # 1:10
  pos(rowRanges(genoset.ds))  # The same
}
