% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbindDataframe}
\alias{rbindDataframe}
\title{A fast method for concatenating data.frames}
\usage{
rbindDataframe(dflist, element.colname)
}
\arguments{
\item{dflist}{list of data.frames}

\item{element.colname}{scalar character, name for additional factor column giving the name
of the element of `dflist` corresponding to each row. `dflist` must be named to use this feature.}
}
\value{
data.frame
}
\description{
Performs the same action as do.call(rbind, list_of_dataframes), but dramatically faster. Part
of the speed comes from assuming that all of the data.frames have the same column names and
types. If desirved an additional factor column can be added that specifies the original list
element associated with each row. The argument `element.colname` is used to
}

