\name{featureNames<-}
\alias{featureNames<-}
\alias{featureNames<-,GenoSet-method}
\alias{featureNames<-,GRanges-method}
\alias{featureNames<-,RangedData-method}
\title{Set featureNames}
\arguments{
  \item{object}{GenoSet, RangedData, or GRanges}

  \item{value}{ANY}
}
\value{
  A new object of the class of supplied object
}
\description{
  Set featureNames
}
\details{
  Set featureNames of a GenoSet, GRanges, or RangedData
  (rownames, names, or rownames respectively).
}
\author{
  Peter M. Haverty
}

