% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{show,GenoSet-method}
\alias{show,GenoSet-method}
\title{Print a GenoSet}
\usage{
\S4method{show}{GenoSet}(object)
}
\arguments{
  \item{object}{a GenoSet}
}
\description{
Prints out a description of a GenoSet object
}

