% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{universe,GenoSet-method}
\alias{universe,GRanges-method}
\alias{universe,GenoSet-method}
\alias{universe<-,GRanges-method}
\alias{universe<-,GenoSet-method}
\title{Get universe annotations}
\usage{
\S4method{universe}{GenoSet}(x)

\S4method{universe}{GRanges}(x)

\S4method{universe}{GenoSet}(x) <- value

\S4method{universe}{GRanges}(x) <- value
}
\arguments{
  \item{x}{GenoSet}

  \item{value}{scalar character, new value of universe}
}
\value{
scalar character
}
\description{
Get universe annotations
}

