% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{width,GenoSet-method}
\alias{width,GenoSet-method}
\title{Get width of location for each feature}
\usage{
\S4method{width}{GenoSet}(x)
}
\arguments{
  \item{x}{GenoSet}
}
\value{
integer
}
\description{
Get width of location for each feature
}

