% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk.R
\name{API}
\alias{API}
\title{Converting eSet API to RangedSummarizedExperiment}
\arguments{
\item{x}{GenoSet}

\item{object}{GenoSet}

\item{i}{name or index of assay to select}

\item{value}{new value to insert in assays, a matrix-like object}

\item{...}{extra arguments}

\item{withDimnames}{scalar logical, should assays be returned with dimnames? Currently always TRUE for GenoSet.}
}

