\name{assayDataElement}
\alias{assayDataElement}
\title{Get assayDataElement, attaching on-disk resource if necessary}
\usage{
  assayDataElement(object, elt)
}
\arguments{
  \item{object}{eSet}

  \item{elt}{character}
}
\value{
  assayDataElement, matrix, DataFrame, or the like
}
\description{
  Get assayDataElement, attaching on-disk resource if
  necessary
}
\author{
  Peter M. Haverty \email{phaverty@gene.com}
}

