\docType{methods}
\name{baf}
\alias{baf}
\alias{baf,BAFSet-method}
\alias{baf<-}
\alias{baf<-,BAFSet,matrix-method}
\title{Get baf data}
\usage{
baf(object)

\S4method{baf}{BAFSet}(object)

baf(object) <- value

\S4method{baf}{BAFSet,matrix}(object) <- value
}
\arguments{
  \item{object}{A BAFset object}
}
\value{
matrix
}
\description{
Get or Set the baf assayData slot
}
\details{
Get or Set the baf assayData slot
}
\author{
Peter M. Haverty
}

