\name{baf}
\alias{baf-methods}
\alias{baf}
\title{Get or Set the baf assayData slot...}
\description{Get or Set the baf assayData slot}
\details{\code{baf-methods}: Get or Set the baf assayData slot

}
\value{\code{baf-methods}: matrix

}
\author{Peter M. Haverty}
\arguments{\item{object}{A BAFset object}
}
\examples{data(genoset)
baf(baf.ds)  # Returns assayDataElement called "baf"
baf(baf.ds) <- baf2mbaf( baf(baf.ds) )}
\alias{baf,-method}
\alias{baf}
\alias{baf<--methods}
\alias{baf<-}
\alias{baf<-,-method}
\alias{baf<-}

