% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{chr}
\alias{chr}
\alias{chr,GenoSet-method}
\alias{chr,GenomicRanges-method}
\title{Chromosome name for each feature}
\usage{
chr(object)

\S4method{chr}{GenoSet}(object)

\S4method{chr}{GenomicRanges}(object)
}
\arguments{
  \item{object}{GRanges GenoSet}
}
\value{
character vector of chromosome positions for each feature
}
\description{
Get chromosome name for each feature. Returns character.
}
\examples{
data(genoset)
  chr(genoset.ds)  # c("chr1","chr1","chr1","chr1","chr3","chr3","chrX","chrX","chrX","chrX")
  chr(locData(genoset.ds))  # The same
}

