% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{chr}
\alias{chr}
\alias{chr,GRanges-method}
\alias{chr,GenoSet-method}
\title{Chromosome name for each feature}
\usage{
chr(object)

\S4method{chr}{GenoSet}(object)

\S4method{chr}{GRanges}(object)
}
\arguments{
  \item{object}{GRanges GenoSet}
}
\value{
character vector of chromosome positions for each feature
}
\description{
Get chromosome name for each feature.  Returns character.
}
\examples{
data(genoset)
  chr(genoset.ds)  # c("chr1","chr1","chr1","chr1","chr3","chr3","chrX","chrX","chrX","chrX")
  chr(locData(genoset.ds))  # The same
}

