% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{chrIndices}
\alias{chrIndices}
\alias{chrIndices,GenoSetOrGenomicRanges-method}
\title{Get a matrix of first and last index of features in each chromosome}
\usage{
chrIndices(object, chr = NULL)

\S4method{chrIndices}{GenoSetOrGenomicRanges}(object, chr = NULL)
}
\arguments{
  \item{object}{GenoSet or GRanges}

  \item{chr}{character, specific chromosome name}
}
\value{
data.frame with "first" and "last" columns
}
\description{
Sometimes it is handy to know the first and last index for
each chr. This is like chrInfo but for feature indices
rather than chromosome locations. If chr is specified, the
function will return a sequence of integers representing
the row indices of features on that chromosome.
}
\examples{
data(genoset)
  chrIndices(genoset.ds)
  chrIndices(locData(genoset.ds))  # The same
}

