% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{chrInfo}
\alias{chrInfo}
\alias{chrInfo,GenoSetOrGenomicRanges-method}
\title{Get chromosome start and stop positions}
\usage{
chrInfo(object)

\S4method{chrInfo}{GenoSetOrGenomicRanges}(object)
}
\arguments{
  \item{object}{A GenoSet object or similar}
}
\value{
list with start and stop position, by ordered chr
}
\description{
Provides a matrix of start, stop and offset, in base
numbers for each chromosome.
}
\examples{
data(genoset)
  chrInfo(genoset.ds)
  chrInfo(locData(genoset.ds))  # The same
}

