% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{chrNames}
\alias{chrNames}
\alias{chrNames,GenoSet-method}
\alias{chrNames,GenomicRanges-method}
\alias{chrNames<-}
\alias{chrNames<-,GenoSet-method}
\alias{chrNames<-,GenomicRanges-method}
\title{Get list of unique chromosome names}
\usage{
chrNames(object)

\S4method{chrNames}{GenoSet}(object)

\S4method{chrNames}{GenomicRanges}(object)

chrNames(object) <- value

\S4method{chrNames}{GenoSet}(object) <- value

\S4method{chrNames}{GenomicRanges}(object) <- value
}
\arguments{
\item{object}{GenomicRanges or GenoSet}

\item{value}{return value of chrNames}
}
\value{
character vector with names of chromosomes
}
\description{
Get list of unique chromosome names
}
\examples{
  data(genoset)
  chrNames(genoset.ds) # c("chr1","chr3","chrX")
  chrNames(locData(genoset.ds))  # The same
  chrNames(genoset.ds) = sub("^chr","",chrNames(genoset.ds))
}
