% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ordering.R
\name{chrOrder}
\alias{chrOrder}
\title{Order chromosome names in proper genome order}
\usage{
chrOrder(chr.names)
}
\arguments{
  \item{chr.names}{character, vector of unique chromosome
  names}
}
\value{
character vector of chromosome names in proper order
}
\description{
Chromosomes make the most sense orded by number, then by
letter.
}
\examples{
chrOrder(c("chr5","chrX","chr3","chr7","chrY"))  #  c("chr3","chr5","chr7","chrX","chrY")
}
\seealso{
Other "genome ordering": \code{\link{isGenomeOrder}},
\code{\link{isGenomeOrder,GRanges-
method}}, \code{\link{isGenomeOrder,GenoSet-
method}}; \code{\link{toGenomeOrder}},
\code{\link{toGenomeOrder,GRanges-method}},
\code{\link{toGenomeOrder,GenoSet-method}}
}

