\docType{methods}
\name{cn}
\alias{cn}
\alias{cn,CNSet-method}
\alias{cn<-}
\alias{cn<-,CNSet,matrix-method}
\title{Get or Set the cn assayData slot}
\usage{
cn(object)

\S4method{cn}{CNSet}(object)

cn(object) <- value

\S4method{cn}{CNSet,matrix}(object) <- value
}
\arguments{
  \item{object}{A BAFset object}
}
\value{
matrix
}
\description{
Get or Set the cn assayData slot
}
\author{
Peter M. Haverty
}

