\name{cn}
\alias{cn-methods}
\alias{cn}
\title{Get or Set the cn assayData slot...}
\description{Get or Set the cn assayData slot}
\details{\code{cn-methods}: Get or Set the cn assayData slot

}
\value{\code{cn-methods}: matrix

}
\author{Peter M. Haverty}
\arguments{\item{object}{A BAFset object}
}
\examples{data(genoset)
cn(cn.ds)  # Returns assayDataElement called "cn"
cn(cn.ds) <- cn(cn.ds) + 5}
\alias{cn,-method}
\alias{cn}
\alias{cn<--methods}
\alias{cn<-}
\alias{cn<-,-method}
\alias{cn<-}

