% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{colnames,GenoSet-method}
\alias{colnames,GenoSet-method}
\alias{colnames<-,GenoSet-method}
\alias{sampleNames,GenoSet-method}
\alias{sampleNames<-,GenoSet,ANY-method}
\title{Get colnames from a GenoSet}
\usage{
\S4method{colnames}{GenoSet}(x)

\S4method{colnames}{GenoSet}(x) <- value

\S4method{sampleNames}{GenoSet}(object)

\S4method{sampleNames}{GenoSet,ANY}(object) <- value
}
\arguments{
  \item{x}{GenoSet}

  \item{value}{a valid value for that component of
  \code{\link{dimnames}(x)}.  For a matrix or array this is
  either \code{NULL} or a character vector of non-zero
  length equal to the appropriate dimension.}

  \item{object}{Object, possibly derived from class
  \code{eSet}.}
}
\value{
character vector with names of samples
}
\description{
Get colnames from a GenoSet
}
\examples{
data(genoset)
  head(colnames(genoset.ds))
}

