% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{colnames,GenoSet-method}
\alias{colnames,GenoSet-method}
\alias{colnames<-,GenoSet-method}
\alias{sampleNames,GenoSet-method}
\alias{sampleNames<-,GenoSet,ANY-method}
\title{Get colnames from a GenoSet}
\usage{
\S4method{colnames}{GenoSet}(x)

\S4method{colnames}{GenoSet}(x) <- value

\S4method{sampleNames}{GenoSet}(object)

\S4method{sampleNames}{GenoSet,ANY}(object) <- value
}
\arguments{
  \item{x}{GenoSet}
}
\value{
character vector with names of samples
}
\description{
Get colnames from a GenoSet
}
\examples{
data(genoset)
  head(colnames(genoset.ds))
}

