% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{elementLengths,GenoSet-method}
\alias{elementLengths,GenoSet-method}
\alias{elementLengths,GenomicRanges-method}
\title{Get elementLengths from locData slot}
\usage{
\S4method{elementLengths}{GenoSet}(x)

\S4method{elementLengths}{GenomicRanges}(x)
}
\arguments{
  \item{x}{GenoSet}
}
\value{
character
}
\description{
Get elementLengths from locData slot
}

