\name{featureNames<-}
\alias{featureNames<-}
\title{Set featureNames}
\arguments{
  \item{object}{GenoSet}

  \item{value}{ANY}
}
\value{
  A new object of the class of supplied object
}
\description{
  Set featureNames
}
\details{
  Set featureNames including rownames of position info
}
\author{
  Peter M. Haverty
}

