% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{genome,GenoSet-method}
\alias{genome,GenoSet-method}
\alias{genome<-,GenoSet-method}
\title{Genome version}
\usage{
\S4method{genome}{GenoSet}(x)

\S4method{genome}{GenoSet}(x) <- value
}
\arguments{
  \item{x}{GenoSet}
}
\value{
character, e.g. hg19
}
\description{
The genome positions of the features in locData. The UCSC
notation (e.g. hg18, hg19, etc.) should be used.
}
\examples{
data(genoset)
  genome(genoset.ds)
  genome(genoset.ds) = "hg19"
}

