% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{genome,GenoSet-method}
\alias{genome,GenoSet-method}
\alias{genome<-,GenoSet-method}
\title{Genome version}
\usage{
\S4method{genome}{GenoSet}(x)

\S4method{genome}{GenoSet}(x) <- value
}
\arguments{
  \item{x}{GenoSet}

  \item{value}{scalar character, incoming genome string}
}
\value{
character, e.g. hg19
}
\description{
The genome positions of the features in locData. The UCSC
notation (e.g. hg18, hg19, etc.) should be used.
}
\examples{
data(genoset)
  genome(genoset.ds)
  genome(genoset.ds) = "hg19"
}

