% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{genoPos}
\alias{genoPos}
\alias{genoPos,GenoSetOrGenomicRanges-method}
\title{Get base positions of features in genome-scale units}
\usage{
genoPos(object)

\S4method{genoPos}{GenoSetOrGenomicRanges}(object)
}
\arguments{
\item{object}{A GenoSet object or a GenomicRanges object}
}
\value{
numeric position of each feature in whole genome units, in original order
}
\description{
Get base positions of array features in bases counting from the start of the
genome. Chromosomes are ordered numerically, when possible, then lexically.
}
\examples{
  data(genoset)
  head(genoPos(genoset.ds))
  head(genoPos(locData(genoset.ds)))  # The same
}
