\name{genoset-datasets}
\docType{data}
\alias{genoset-datasets}
\alias{genoset.ds}
\title{Example GenoSet object}
\description{A GenoSet object the 'baf' (B-Allele Frequency) and 'lrr' (Log-R Ratio)
assay matrices. The 'lrr' assay matrix contains DNA copy number on the
scale of tumor/ploidy and the 'baf' assay matrix contains data in the
range 0 to 1 where 0 indicates the AA genotype, 0.5 indicates the AB
genotype and 1 indicates the BB genotype.
}
\source{Simulated data}
\keyword{datasets}
