% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genoset-deprecated}
\alias{genoset-deprecated}
\title{Deprecated genoset features}
\description{
GenoSet is moving towards using GenomicRanges instead of
RangedData. We are also getting rid of dependencies on eSet
for a potential switch to an underlying
SummarizedExperiment.
}

