% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/junk.R
\name{genoset-deprecated}
\alias{genoset-deprecated}
\title{Deprecated genoset features}
\description{
GenoSet is moving towards using GenomicRanges instead of
RangedData. We are also getting rid of dependencies on eSet
for a potential switch to an underlying
SummarizedExperiment.
}

