% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{package}
\name{genoset-package}
\alias{genoset-package}
\alias{genoset}
\title{GenoSet: An eSet for data with genome locations}
\description{
Load, manipulate, and plot copynumber and BAF data.
}
\seealso{
genoset-datasets
}
