% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{isGenomeOrder}
\alias{isGenomeOrder}
\title{Check if a GRanges orGenoSet is in genome order}
\usage{
isGenomeOrder(ds, strict = TRUE)
}
\arguments{
\item{ds}{GenoSet or GRanges}

\item{strict}{logical, should space/chromosome order be identical to that from chrOrder?}
}
\value{
logical
}
\description{
Checks that rows in each chr are ordered by start.  If strict=TRUE, then chromosomes
must be in order specified by chrOrder. isGenomeOrder for GRanges differs from order
in that it orders by chromsome and start position only,
rather than chromsome, strand, start, and width.
}
\examples{
  data(genoset,package='genoset')
  isGenomeOrder( rowRanges(genoset.ds) )
}
\seealso{
Other 'genome ordering': \code{\link{chrOrder}},
  \code{\link{toGenomeOrder}}
}
\concept{'genome ordering'}
