% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{locData}
\alias{locData}
\alias{locData,GenoSet-method}
\alias{locData<-}
\alias{locData<-,GenoSet,GenomicRanges-method}
\title{Access the feature genome position info}
\usage{
locData(object)

\S4method{locData}{GenoSet}(object)

locData(object) <- value

\S4method{locData}{GenoSet,GenomicRanges}(object) <- value
}
\arguments{
\item{object}{GenoSet}

\item{value}{GRanges describing features}
}
\value{
A GenoSet object
}
\description{
The position information for each probe/feature is stored as an GRanges object.
The locData functions allow this data to be accessed or re-set.
}
\examples{
data(genoset)
rd = locData(genoset.ds)
locData(genoset.ds) = rd
}
