% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lr2cn}
\alias{lr2cn}
\title{Take vector or matrix of log2 ratios, convert to copynumber}
\usage{
lr2cn(x)
}
\arguments{
\item{x}{numeric data in log2ratio values}
}
\value{
data of same type as "x" transformed into copynumber units
}
\description{
Utility function for converting log2ratio units (zero is normal) to copynumber units (two is normal)
}
\seealso{
cn2lr
}
