\name{lrr}
\alias{lrr-methods}
\alias{lrr}
\title{Get or Set the lrr assayData slot...}
\description{Get or Set the lrr assayData slot}
\details{\code{lrr-methods}: Get or Set the lrr assayData slot

}
\value{\code{lrr-methods}: matrix

}
\author{Peter M. Haverty}
\arguments{\item{object}{A BAFset object}
}
\examples{data(genoset)
lrr(baf.ds)  # Returns assayDataElement called "lrr"
lrr(baf.ds) <- lrr(baf.ds) + 0.1}
\alias{lrr,-method}
\alias{lrr}
\alias{lrr<--methods}
\alias{lrr<-}
\alias{lrr<-,-method}
\alias{lrr<-}

