\docType{methods}
\name{lrr}
\alias{lrr}
\alias{lrr,BAFSet-method}
\alias{lrr<-}
\alias{lrr<-,BAFSet,matrix-method}
\title{Get lrr data}
\usage{
lrr(object)

\S4method{lrr}{BAFSet}(object)

lrr(object) <- value

\S4method{lrr}{BAFSet,matrix}(object) <- value
}
\arguments{
  \item{object}{A BAFset object}
}
\value{
matrix
}
\description{
Get or Set the lrr assayData slot
}
\details{
Get or Set the lrr assayData slot
}
\author{
Peter M. Haverty
}

