\name{lrr}
\alias{lrr}
\alias{lrr,BAFSet-method}
\alias{lrr<-}
\alias{lrr<-,BAFSet,matrix-method}
\title{Get lrr data}
\arguments{
  \item{object}{A BAFset object}
}
\value{
  matrix
}
\description{
  Get or Set the lrr assayData slot
}
\details{
  Get or Set the lrr assayData slot
}
\author{
  Peter M. Haverty
}

