% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{nrow,GenomicRanges-method}
\alias{nrow,GenomicRanges-method}
\title{Number of rows}
\usage{
\S4method{nrow}{GenomicRanges}(x)
}
\arguments{
  \item{x}{GRanges or GenoSet}
}
\value{
integer
}
\description{
Number of rows
}

