% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{pos}
\alias{pos}
\alias{pos,GenoSetOrGenomicRanges-method}
\title{Chromosome position of features}
\usage{
pos(object)

\S4method{pos}{GenoSetOrGenomicRanges}(object)
}
\arguments{
\item{object}{GRanges GenoSet}
}
\value{
numeric vector of feature positions within a chromosome
}
\description{
Get chromosome position of features/ranges. Defined as floor of mean of start and end.
}
\examples{
  data(genoset)
  pos(genoset.ds)  # 1:10
  pos(locData(genoset.ds))  # The same
}
