% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bounds.R
\name{rangeSampleMeans}
\alias{rangeSampleMeans}
\title{Average features in ranges per sample}
\usage{
rangeSampleMeans(query, subject, assay.element, na.rm = FALSE)
}
\arguments{
  \item{query}{GRanges object representing genomic regions
  (genes) to be averaged.}

  \item{subject}{A GenoSet object or derivative}

  \item{assay.element}{character, name of element in
  assayData to use to extract data}

  \item{na.rm}{scalar logical, ignore NAs?}
}
\value{
numeric matrix of features in each range averaged by sample
}
\description{
This function takes per-feature genomic data and returns
averages for each of a set of genomic ranges. The most
obvious application is determining the copy number of a set
of genes. The features corresponding to each gene are
determined with boundingIndices such that all features with
the bounds of a gene (overlaps). The features on either
side of the gene unless those positions exactly match the
first or last base covered by the gene.  Therefore, genes
falling between two features will at least cover two
features. Range bounding is performed by the
boundingIndices function.
}
\examples{
data(genoset)
  my.genes = GRanges( ranges=IRanges(start=c(35e6,128e6),end=c(37e6,129e6),names=c("HER2","CMYC")), seqnames=c("chr17","chr8") )
  rangeSampleMeans( my.genes, genoset.ds, "lrr" )
}
\seealso{
Other "range summaries":
\code{\link{boundingIndicesByChr}};
\code{\link{boundingIndices}}
}

