% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{rangeSegMeanLength}
\alias{rangeSegMeanLength}
\alias{rangeSegMeanLength,GRanges,list-method}
\alias{rangeSegMeanLength,GRanges,data.frame-method}
\title{Get segment widths}
\usage{
rangeSegMeanLength(range.gr, segs)

\S4method{rangeSegMeanLength}{GRanges,list}(range.gr, segs)

\S4method{rangeSegMeanLength}{GRanges,data.frame}(range.gr, segs)
}
\arguments{
\item{range.gr}{GRanges, genome regions of interest, usually genes}

\item{segs}{data.frame of segments, like from segTable, or a list of these}
}
\value{
named vector of lengths, one per item in range.gr, or a range x length(segs) of these if segs is also list-like.
}
\description{
The width of a genomic segment helps inform us about the importance of a copy number value. Focal amplifications
are more interesting than broad gains, for example. Given a range of interesting regions (i.e. genes) this
function determines all genomics segments covered by each gene and returns the average length of the
segments covered by each gene in each sample. Often only a single segment covers a given gene in
a given sample.
}
\seealso{
Other "segmented data": 
\code{\link{bounds2Rle}()},
\code{\link{runCBS}()},
\code{\link{segPairTable}()},
\code{\link{segTable}()},
\code{\link{segs2Granges}()},
\code{\link{segs2RleDataFrame}()},
\code{\link{segs2Rle}()}
}
\concept{"segmented data"}
