% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readGenoSet}
\alias{readGenoSet}
\title{Load a GenoSet from a RData file}
\usage{
readGenoSet(path)
}
\arguments{
\item{path}{character, path to rds or rda file}
}
\value{
GenoSet or related object (only object in RData file)
}
\description{
Given a rds file or a rda file with one GenoSet, load it,
and return. Objects that pre-date the switch to a RangedSummarizedExperiment internal representation (V
1.29.0) are automatically switched to the new format.
}
\examples{
\dontrun{ ds = readGenoSet('/path/to/genoset.RData') }
\dontrun{ ds = readGenoSet('/path/to/genoset.rda') }
\dontrun{ ds = readGenoSet('/path/to/genoset.rds') }
}
