% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoset-class.R
\docType{methods}
\name{rownames,GenomicRanges-method}
\alias{featureNames,GenoSet-method}
\alias{featureNames,GenomicRanges-method}
\alias{featureNames<-,GenoSet-method}
\alias{featureNames<-,GenomicRanges-method}
\alias{rownames,GenoSet-method}
\alias{rownames,GenomicRanges-method}
\alias{rownames<-,GenoSet-method}
\alias{rownames<-,GenomicRanges-method}
\title{Get rownames from GRanges, or GenoSet}
\usage{
\S4method{rownames}{GenomicRanges}(x)

\S4method{rownames}{GenoSet}(x)

\S4method{rownames}{GenomicRanges}(x) <- value

\S4method{rownames}{GenoSet}(x) <- value

\S4method{featureNames}{GenoSet}(object)

\S4method{featureNames}{GenomicRanges}(object)

\S4method{featureNames}{GenoSet}(object) <- value

\S4method{featureNames}{GenomicRanges}(object) <- value
}
\arguments{
  \item{x}{GRanges or GenoSet} \item{value}{character,
  incoming rownames} \item{object}{GenoSet}
}
\value{
character vector with names rows/features
}
\description{
Get rownames from GRanges or GenoSet.
}
\examples{
data(genoset)
  head(rownames(locData.gr))
  head(rownames(genoset.ds))
}

