% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{rownames,GRanges-method}
\alias{featureNames,GRanges-method}
\alias{featureNames,GenoSet-method}
\alias{featureNames<-,GRanges-method}
\alias{featureNames<-,GenoSet-method}
\alias{rownames,GRanges-method}
\alias{rownames,GenoSet-method}
\alias{rownames<-,GRanges-method}
\alias{rownames<-,GenoSet-method}
\title{Get rownames from GRanges, or GenoSet}
\usage{
\S4method{rownames}{GRanges}(x)

\S4method{rownames}{GenoSet}(x)

\S4method{rownames}{GRanges}(x) <- value

\S4method{rownames}{GenoSet}(x) <- value

\S4method{featureNames}{GenoSet}(object)

\S4method{featureNames}{GRanges}(object)

\S4method{featureNames}{GenoSet}(object) <- value

\S4method{featureNames}{GRanges}(object) <- value
}
\arguments{
  \item{object}{GRanges or GenoSet}
}
\value{
character vector with names rows/features
}
\description{
Get rownames from GRanges or GenoSet.
}
\examples{
data(genoset)
  head(rownames(locData.gr))
  head(rownames(genoset.ds))
}

