% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/segments.R
\name{segs2Rle}
\alias{segs2Rle}
\title{Make Rle from segments for one sample}
\usage{
segs2Rle(segs, locs)
}
\arguments{
  \item{segs}{data.frame of segments, formatted as output
  of segment function from DNAcopy package}

  \item{locs}{GenomicRanges, like locData slot of a
  GenoSet}
}
\value{
Rle with run lengths and run values covering all features
in the data set.
}
\description{
Take output of CBS, make Rle representing all features in
'locs' ranges. CBS output contains run length and run
values for genomic segmetns, which could very directly be
converted into a Rle. However, as NA values are often
removed, especially for mBAF data, these run lengths do not
necessarily cover all features in every sample. Using the
start and top positions of each segment and the location of
each feature, we can make a Rle that represents all
features.
}
\examples{
data(genoset)
  segs = runCBS( genoset.ds[, , "lrr"], locData(genoset.ds), return.segs=TRUE )
  segs2Rle( segs[[1]], locData(genoset.ds) )  # Take a data.frame of segments, say from DNAcopy's segment function, and make Rle's using probe locations in the locs
}
\seealso{
Other "segmented data": \code{\link{bounds2Rle}};
\code{\link{rangeSegMeanLength}},
\code{\link{rangeSegMeanLength,GRanges,data.frame-method}},
\code{\link{rangeSegMeanLength,GRanges,list-method}};
\code{\link{runCBS}}; \code{\link{segPairTable}},
\code{\link{segPairTable,DataFrame,DataFrame-method}},
\code{\link{segPairTable,Rle,Rle-method}};
\code{\link{segTable}},
\code{\link{segTable,DataFrame-method}},
\code{\link{segTable,Rle-method}};
\code{\link{segs2Granges}}; \code{\link{segs2RleDataFrame}}
}

