% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segs2RleDataFrame}
\alias{segs2RleDataFrame}
\title{CBS segments to probe matrix}
\usage{
segs2RleDataFrame(seg.list, locs)
}
\arguments{
\item{seg.list}{list, list of data frames, one per sample, each is result from CBS}

\item{locs}{locData from a GenoSet object}
}
\value{
RleDataFrame with nrows same as locs and one column for each sample
}
\description{
Given segments, make an RleDataFrame of Rle objects for each sample
}
\details{
Take table of segments from CBS, convert DataTable of Rle objects for each sample.
}
\examples{
  data(genoset)
  seg.list = runCBS( genoset.ds[, , "lrr"], locData(genoset.ds), return.segs=TRUE )
  segs2RleDataFrame( seg.list, locData(genoset.ds) )  # Loop segs2Rle on list of data.frames in seg.list
}
\seealso{
Other "segmented data": 
\code{\link{bounds2Rle}()},
\code{\link{rangeSegMeanLength}()},
\code{\link{runCBS}()},
\code{\link{segPairTable}()},
\code{\link{segTable}()},
\code{\link{segs2Granges}()},
\code{\link{segs2Rle}()}
}
\concept{"segmented data"}
\concept{segments}
