\name{start}
\alias{dim}
\alias{elementLengths}
\alias{end}
\alias{names}
\alias{nrow}
\alias{start}
\alias{width}
\title{Get start of location for each feature}
\arguments{
  \item{x}{GenoSet}

  \item{x}{GenoSet}

  \item{x}{GenoSet}

  \item{x}{GenoSet}

  \item{x}{GenoSet}
}
\value{
  integer

  integer

  integer

  character

  character
}
\description{
  Get start of location for each feature

  Get end of location for each feature

  Get width of location for each feature

  Get names of data matrices. For the time being, this is
  \code{assayDataElementNames}. This function used to do
  \code{chrNames}.

  Get elementLengths from locData slot
}
\details{
  Get elementLengths from locData slot
}

