% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{start,GenoSet-method}
\alias{start,GenoSet-method}
\title{Get start of location for each feature}
\usage{
\S4method{start}{GenoSet}(x)
}
\arguments{
  \item{x}{GenoSet}
}
\value{
integer
}
\description{
Get start of location for each feature
}

