% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{subsetAssayData}
\alias{subsetAssayData}
\title{Subset or re-order assayData}
\usage{
subsetAssayData(orig, i, j, ..., drop = FALSE)
}
\arguments{
  \item{orig}{assayData environment} \item{i}{row indices}
  \item{j}{col indices} \item{...}{Additional args to give to
  subset operator} \item{drop}{logical, drop dimensions when
  subsetting with single value?}
}
\value{
assayData data structure
}
\description{
Subset or re-order assayData locked environment,
environment, or list. Shamelessly stolen from "["
method in Biobase version 2.8 along with guts of
assayDataStorageMode()
}
\examples{
data(genoset)
  ad = assayData(genoset.ds)
  small.ad = subsetAssayData(ad,1:5,2:3)
}

