% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoset-class.R
\name{width,GenoSet-method}
\alias{width,GenoSet-method}
\title{Get width of location for each feature}
\usage{
\S4method{width}{GenoSet}(x)
}
\arguments{
\item{x}{GenoSet}
}
\value{
integer
}
\description{
Get width of location for each feature
}
